requirements.txt
Django==3.0.8
gunicorn==20.0.4
djangorestframework
django-cors-headers==3.5.0
gunicorn
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
